global-news-tracker/requirements.txt
# Backend
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
SQLAlchemy>=2.0,<3.0
pandas==2.1.1
python-dotenv==1.0.0
scikit-learn>=1.0
openai>=1.0.0
httpx>=0.27.0
transformers>=4.40.0
torch>=2.1.0
spacy>=3.6.0

# Frontend
streamlit==1.28.1
plotly==5.18.0
matplotlib==3.8.0

